add dashboard page
